laba2 python side added

+++ Laba1/Task3.cs
@@ -14,7 +14,7 @@ namespace PZPK_Labas.Laba1
 
         private static void GetFunctionValue()
         {
-            int x;
+            double x;
             ConsoleInput.SafeIntInput("Введіть значення аргументу", out x);
 
             if (x < -4 || x > 10)
